refactored component directories
