+++ components/Home/AdminHomePage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import { AdminCalendar } from "../Calendar/AdminCalendar";
-import { CreateEventsWizard } from "../CreateEventsWizard/CreateEventsWizard";
+import { AdminCalendar } from "./Calendar/AdminCalendar";
+import { CreateClassWizard } from "./CreateClassWizard/CreateClassWizard";
 import styles from "./AdminHomePage.module.css";
 
 const AdminHomePage: React.FC<object> = () => {
@@ -16,7 +16,7 @@ const AdminHomePage: React.FC<object> = () => {
           Create
           <img className={styles.addIcon} src="AddIcon.png" />
         </button>
-        {showWizard ? <CreateEventsWizard handleClose={handleClose} /> : null}
+        {showWizard ? <CreateClassWizard handleClose={handleClose} /> : null}
       </div>
       <AdminCalendar />;
     </div>
