+++ components/League/ClassCard.tsx
