+++ components/League/StudentCard.tsx
