+++ components/League/StudentView.tsx
