+++ pages/league.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import styles from "../styles/League.module.css";
-import { ClassView } from "../components/LeaguePage/ClassView";
-import { StudentView } from "../components/LeaguePage/StudentView";
-import { StaffView } from "../components/LeaguePage/StaffView";
+import { ClassView } from "../components/League/ClassView";
+import { StudentView } from "../components/League/StudentView";
+import { StaffView } from "../components/League/StaffView";
 import { Class } from "../models/class";
 import { User } from "../models/users";
 import { Student } from "../models/students";
